requirements.txt
absl-py==2.1.0
blinker==1.4
certifi==2024.6.2
charset-normalizer==3.3.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.6.1
environs==9.5.0
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
grpcio==1.56.0
grpcio-tools==1.37.1
httplib2==0.20.2
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==4.6.4
imutils==0.5.4
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
milvus==2.3.5
mmh3==3.0.0
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==3.20.3
PyGObject==3.42.1
PyJWT==2.3.0
pymilvus==2.3.0
pymongo==4.7.3
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
safetensors==0.4.3
scipy==1.14.0
SecretStorage==3.3.1
six==1.16.0
sympy==1.12.1
systemd-python==234
tensorboard==2.17.0
tensorboard-data-server==0.7.2
timm==1.0.7
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.3
zipp==1.0.0
